% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-curation.R
\name{make.ancestor}
\alias{make.ancestor}
\title{Reconstruct an ancestral state based on two specified daughters}
\usage{
make.ancestor(s1, s2)
}
\arguments{
\item{s1}{A binary string, possibly with "?"s}

\item{s2}{A binary string, possibly with "?"s}
}
\value{
A binary string describing the ancestor state
}
\description{
Data is specified as binary strings of equal length, which may include "?" for uncertainty
}
\examples{
make.ancestor("101", "0??")
}
