% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HyperLAU}
\alias{HyperLAU}
\title{HyperLAU computation (Armadillo)}
\usage{
HyperLAU(obs, initialstates = NULL, nboot = 0L, model = -1L, rate = 1.001)
}
\arguments{
\item{obs}{Numeric matrix of observations}

\item{initialstates}{optional}

\item{nboot}{optional}

\item{model}{optional}

\item{rate}{optional}
}
\value{
A fitted HyperLAU model, including a dataframe of transitions, vector of likelihoods, and the dimensionality of the problem
}
\description{
HyperLAU computation (Armadillo)
}
\examples{
data = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = HyperLAU(data)
}
